move model_run() tests to test_ogs6py.py
